Fix mismatched types.

+++ src/bit_iterator.rs
@@ -18,10 +18,10 @@ impl<'a> Iterator for BitIterator<'a> {
     fn next(&mut self) -> Option<u8> {
         let result = self.object[self.current_byte] >> self.current_bit & 0x1;
 
-        if self.current_bit + 1 == ::std::u8::BITS {
+        if self.current_bit + 1 == ::std::u8::BITS as usize {
             self.current_byte += 1;
         }
-        self.current_bit = (self.current_bit + 1) % ::std::u8::BITS;
+        self.current_bit = (self.current_bit + 1) % ::std::u8::BITS as usize;
 
         if self.current_byte == self.object.len() {
             return None;
